+++ tools/explorerGambits/gambits.js
@@ -21709,6 +21709,26 @@
         }
       ]
     },
+    "rnbqk2rppp2ppp5n23P41b2p32PP1N2PP2QPPPRNB1KB1Rb": {
+      "total": 1,
+      "moves": [
+        {
+          "uci": "e8h8",
+          "san": "O-O",
+          "nr": 1
+        }
+      ]
+    },
+    "rnbqk2rppp2ppp5n23P41b2p33P1N2PPP1QPPPRNB1KB1Rw": {
+      "total": 1,
+      "moves": [
+        {
+          "uci": "c2c3",
+          "san": "c3",
+          "nr": 1
+        }
+      ]
+    },
     "rnbqk2rppp2ppp83pP31b2n32NP4PPP3PPR1BQKBNRw": {
       "total": 1,
       "moves": [
@@ -23295,6 +23315,26 @@
         {
           "uci": "f8c5",
           "san": "Bc5",
+          "nr": 1
+        }
+      ]
+    },
+    "rnbqkb1rppp2ppp5n23P44p33P1N2PPP1QPPPRNB1KB1Rb": {
+      "total": 1,
+      "moves": [
+        {
+          "uci": "f8b4",
+          "san": "Bb4+",
+          "nr": 1
+        }
+      ]
+    },
+    "rnbqkb1rppp2ppp5n23P44p35N2PPPPQPPPRNB1KB1Rw": {
+      "total": 1,
+      "moves": [
+        {
+          "uci": "d2d3",
+          "san": "d3",
           "nr": 1
         }
       ]
@@ -27064,6 +27104,26 @@
         }
       ]
     },
+    "rnbqkbnrppp2ppp83P44p35N2PPPP1PPPRNBQKB1Rw": {
+      "total": 1,
+      "moves": [
+        {
+          "uci": "d1e2",
+          "san": "Qe2",
+          "nr": 1
+        }
+      ]
+    },
+    "rnbqkbnrppp2ppp83P44p35N2PPPPQPPPRNB1KB1Rb": {
+      "total": 1,
+      "moves": [
+        {
+          "uci": "g8f6",
+          "san": "Nf6",
+          "nr": 1
+        }
+      ]
+    },
     "rnbqkbnrppp2ppp83P44pP22N5PPPP2PPR1BQKBNRb": {
       "total": 1,
       "moves": [
@@ -27190,17 +27250,17 @@
       ]
     },
     "rnbqkbnrppp2ppp83pp34P35N2PPPP1PPPRNBQKB1Rw": {
-      "total": 7,
+      "total": 8,
       "moves": [
+        {
+          "uci": "e4d5",
+          "san": "exd5",
+          "nr": 3
+        },
         {
           "uci": "f3e5",
           "san": "Nxe5",
           "nr": 5
-        },
-        {
-          "uci": "e4d5",
-          "san": "exd5",
-          "nr": 2
         }
       ]
     },
@@ -27300,16 +27360,16 @@
       ]
     },
     "rnbqkbnrppp2ppp83Pp385N2PPPP1PPPRNBQKB1Rb": {
-      "total": 2,
+      "total": 3,
       "moves": [
-        {
-          "uci": "f8d6",
-          "san": "Bd6",
-          "nr": 1
-        },
         {
           "uci": "e5e4",
           "san": "e4",
+          "nr": 2
+        },
+        {
+          "uci": "f8d6",
+          "san": "Bd6",
           "nr": 1
         }
       ]
@@ -27870,12 +27930,12 @@
       ]
     },
     "rnbqkbnrpppp1ppp84p34P35N2PPPP1PPPRNBQKB1Rb": {
-      "total": 64,
+      "total": 65,
       "moves": [
         {
           "uci": "d7d5",
           "san": "d5",
-          "nr": 7
+          "nr": 8
         },
         {
           "uci": "b8c6",
@@ -27920,12 +27980,12 @@
       ]
     },
     "rnbqkbnrpppp1ppp84p34P38PPPP1PPPRNBQKBNRw": {
-      "total": 128,
+      "total": 129,
       "moves": [
         {
           "uci": "g1f3",
           "san": "Nf3",
-          "nr": 64
+          "nr": 65
         },
         {
           "uci": "f1c4",
@@ -28770,17 +28830,17 @@
       ]
     },
     "rnbqkbnrpppppppp884P38PPPP1PPPRNBQKBNRb": {
-      "total": 211,
+      "total": 212,
       "moves": [
+        {
+          "uci": "e7e5",
+          "san": "e5",
+          "nr": 129
+        },
         {
           "uci": "c7c5",
           "san": "c5",
           "nr": 15
-        },
-        {
-          "uci": "e7e5",
-          "san": "e5",
-          "nr": 128
         },
         {
           "uci": "d7d5",
@@ -28995,17 +29055,17 @@
       ]
     },
     "rnbqkbnrpppppppp8888PPPPPPPPRNBQKBNRw": {
-      "total": 380,
+      "total": 381,
       "moves": [
+        {
+          "uci": "e2e4",
+          "san": "e4",
+          "nr": 212
+        },
         {
           "uci": "d2d4",
           "san": "d4",
           "nr": 105
-        },
-        {
-          "uci": "e2e4",
-          "san": "e4",
-          "nr": 211
         },
         {
           "uci": "g1f3",