+++ tools/forkBehavior/tool.js
@@ -164,6 +164,7 @@
       });
       selectElem.on('change',highlight);
       selectElem.find('option')
+        .on('mousedown',(ev)=>{ if (ev.button==0 && ev.target?.selected) makeMove(); })
         .on('dblclick',makeMove)
         .on('contextmenu',makeMove);
       selectElem.find('li')
@@ -198,6 +199,12 @@
       if (!analysis) return;
       if (analysis.gamebookPlay() || parent.isGamePlaying()) return;
       if (['hybrid','chessbase'].includes(this.options.value)) {
+        $('div.analyse__fork').each((i,e)=>{
+          if (e.lichessTools_forkBehavior_init) return;
+          e.lichessTools_forkBehavior_init=true;
+          e.addEventListener('click',()=>this.inForkClick=true,{ capture: true });
+          e.addEventListener('click',()=>this.inForkClick=false,{ capture: false });
+        });
         if (!parent.isWrappedFunction(analysis.fork.proceed,'forkBehavior')) {
           analysis.fork.proceed=parent.wrapFunction(analysis.fork.proceed,{
             id: 'forkBehavior',
@@ -209,7 +216,7 @@
               const nextSansCount=noDuplicates
                 ? new Set(nextSans).size
                 : nextSans.length;
-              if (nextSansCount>1) {
+              if (!this.inForkClick && nextSansCount>1) {
                 switch(this.options.value) {
                   case 'hybrid':
                     if (this.variationSelect!=analysis.path) {