+++ HISTORY.md
@@ -1,5 +1,12 @@
 History of features added to LiChess Tools in time (reversed order)
 
+17 Jun
+  - extract FENs in PGN Editor
+
+06 Jun
+  - paste images in chat/forum
+  - remove annoying chat warning
+
 04 Jun
   - General tool for extra notifications
   - Add to team notification